Extracting WithGbbHtmlRendering
