+++ lib/stones-spec.rb
@@ -5,6 +5,7 @@ require 'stones'
 
 require_relative 'stones-spec/with_tempfile'
 require_relative 'stones-spec/with_command_line'
+require_relative 'stones-spec/with_gbb_html_rendering'
 require_relative 'stones-spec/version'
 
 require_relative 'stones-spec/language'
